Cleanup class files.
