+++ app_dart/test/model/build_pubsub_test.dart
@@ -1,3 +1,97 @@
+import 'dart:convert';
+
+import 'package:cocoon_service/src/model/luci/build_pubsub.dart';
+import 'package:test/test.dart';
+
+import 'build_pubsub_test_data.dart';
+
 void main() {
-  
+  test('Agent', () {
+    final Agent agent = Agent.fromJson(jsonDecode(agentJson));
+    expect(agent, isNotNull);
+
+    expect(agent.input, isNotNull);
+    expect(agent.input!.data, isNotNull);
+    expect(agent.input!.data!.length, 4);
+    expect(agent.input!.data!.entries.any((element) => element.key == 'bbagent_utility_packages'), isTrue);
+    final InputDataRef inputDataRef = agent.input!.data!.entries.firstWhere((element) => element.key == 'bbagent_utility_packages').value;
+    expect(inputDataRef.cipd!.server == 'chrome-infra-packages.appspot.com', isTrue);
+    expect(inputDataRef.cipd!.specs!.length, 1);
+    expect(inputDataRef.cipd!.specs![0].package == 'infra/tools/luci/cas/platform', isTrue);
+    expect(inputDataRef.cipd!.specs![0].version == 'git_revision:fe9985447e6b95f4907774f05e9774f031700775', isTrue);
+
+    expect(agent.source, isNotNull);
+    expect(agent.source!.cipd, isNotNull);
+    expect(agent.source!.cipd!.package, 'infra/tools/luci/bbagent/platform');
+    expect(agent.source!.cipd!.version, 'latest');
+    expect(agent.source!.cipd!.server, 'chrome-infra-packages.appspot.com');
+
+    expect(agent.purposes, isNotNull);
+    expect(agent.purposes!.entries.any((element) => element.key == 'bbagent_utility_packages' && element.value.name == 'purposeBbAgentUtility'), isTrue);
+    expect(agent.purposes!.entries.any((element) => element.key == 'kitchen-checkout' && element.value.name == 'purposeExePayload'), isTrue);
+
+    expect(jsonEncode(agent.toJson()) == stripJson(agentJson), isTrue);
+  });
+
+  test('BuildBucketV2', () {
+    final BuildBucket buildBucket = BuildBucket.fromJson(jsonDecode(buildBucketV2Json));
+    expect(buildBucket.requestedProperties!.isEmpty, isTrue);
+    expect(buildBucket.hostname! == 'cr-buildbucket-dev.appspot.com', isTrue);
+    expect(buildBucket.experimentReasons!.length, 14);
+    expect(buildBucket.knownPublicGerritHosts!.length, 15);
+    expect(buildBucket.buildNumber, isTrue);
+    expect(buildBucket.agent, isNotNull);
+
+    expect(jsonEncode(buildBucket.toJson()) == stripJson(buildBucketV2Json), isTrue);
+  });
+
+  test('Swarming', () {
+    final Swarming swarming = Swarming.fromJson(jsonDecode(swarmingJson));
+    expect(swarming, isNotNull);
+    expect(swarming.hostname, 'chromium-swarm-dev.appspot.com');
+    expect(swarming.taskId, '62f2e84ef8411d10');
+    expect(swarming.taskServiceAccount, 'chromium-ci-builder-dev@chops-service-accounts.iam.gserviceaccount.com');
+    expect(swarming.priority, 30);
+    expect(swarming.taskDimensions!.length, 3);
+
+    expect(swarming.taskDimensions!.any((element) => element.key == 'pool' && element.value == 'luci.chromium.ci'),
+        isTrue);
+    expect(swarming.taskDimensions!.any((element) => element.key == 'os' && element.value == 'Mac-13'), isTrue);
+    expect(swarming.taskDimensions!.any((element) => element.key == 'cpu' && element.value == 'arm64'), isTrue);
+
+    expect(swarming.caches, isNotNull);
+    expect(swarming.caches!.length, 4);
+    expect(swarming.caches!.any((element) => element.name == 'git' && element.path == 'git'), isTrue);
+    expect(swarming.caches!.any((element) => element.name == 'goma' && element.path == 'goma'), isTrue);
+    expect(
+        swarming.caches!.any((element) =>
+            element.name == 'vpython' && element.path == 'vpython' && element.envVar == 'VPYTHON_VIRTUALENV_ROOT',),
+        isTrue);
+    expect(
+        swarming.caches!.any((element) =>
+            element.name == 'builder_1b2b6e615f25d48545b2db3de147e58b8bea002f690605063288929bc1781d28_v2' &&
+            element.path == 'builder' &&
+            element.waitForWarmCache == '240s',),
+        isTrue,);
+    final String jsonString = jsonEncode(swarming.toJson());
+    assert(jsonString == stripJson(swarmingJson));
+  });
+
+  test('BuildInfra', () {
+    print(buildInfraJson);
+    final BuildInfra buildInfra = BuildInfra.fromJson(jsonDecode(buildInfraJson));
+    expect(buildInfra.buildBucket, isNotNull);
+    expect(buildInfra.swarming, isNotNull);
+    expect(buildInfra.bbAgent, isNotNull);
+    expect(buildInfra.bbAgent!.payloadPath, 'kitchen-checkout');
+    expect(buildInfra.bbAgent!.cacheDir, 'cache');
+  });
+
+  test('Build', () {
+    final Build build = Build.fromJson(jsonDecode(buildJson));
+    expect(build.id, '8777746000874744641');
+    expect(build.builder!.bucket, 'ci');
+    expect(build.builder!.builder, 'mac-arm-rel-dev');
+    expect(build.builder!.project, 'chromium');
+  });
 }
